Insertando multiples combo box en BD y post en empleado Controller.

+++ CursoUdemy/Models/EmpleadoCLS.cs
@@ -40,7 +40,7 @@ namespace CursoUdemy.Models
 
         [Display(Name ="Sueldo")]
         [Required]
-        public double sueldo { get; set; }
+        public decimal sueldo { get; set; }
 
         [Required]
         [Display(Name ="Tipo Contrato")]
